requirements_versions.txt
torchsde==0.2.6
einops==0.8.0
transformers>=4.42.4
ftfy==6.3.1
safetensors==0.4.3
accelerate>=0.32.1
pyyaml==6.0.1
pillow==10.4.0
scipy==1.14.0
tqdm==4.66.4
psutil==6.0.0
pytorch_lightning==2.3.3
omegaconf==2.3.0
gradio>=5.0.0,<6.0.0
pygit2==1.15.1
opencv-contrib-python-headless==4.10.0.84
httpx==0.27.0
onnxruntime>=1.18.1
insightface==0.7.3
timm==1.0.7
numpy>=2.0.0
tokenizers==0.19.1
packaging==24.1
rembg==2.0.57
gguf>=0.9.1
spandrel>=0.4.0
transparent_background==1.2.12
python-dotenv==1.2.2
